databases/sql/requirements.txt
Faker==13.15.1
psycopg2-binary==2.9.3
PyPika==0.48.9
python-dateutil==2.8.2
six==1.16.0
